Apply VS2017 and Qt5 patches (#449)
* Initially applied changes from fork.

* Some fixes to get a bunch of stuff compiling release64. Still not there but getting closer

* Remove unneeded source archives

* Merge forward changes from support/4.11.x

Palette-2458: Remove Opticks n-squared behavior

Palette-2495: Opticks no longer respects extents.
Fixed issue where changing a Point's location didn't update its
PointSet's extents.

Palette-2502: Fix irradiance vs time log plot

Fixed setting of plot scales to log or linear to fully recompute
plot extents, including recomputing the extents of point sets displayed
in the plot.

Fixed setting log/linear axis scales to invalidate the extents cache of
point sets contained in the plot.

Fixes semi-random ordering of plot toolbar buttons by sorting the mouse
mode vector returned by ViewImp::getMouseModes base on mouse mode name
rather than mouse mode pointer.

Changes to support caching of sensor imagery, including optimizing for
case of Can Drop Frames being enabled while animation.

* Some more vs2017 and dependency update fixes.

* All plugins except XmlRpc now build and load

* Fix drag and drop in the layers in the session explorer.

Not sure why Qt wasn't automatically callnig dropMimeData, but it wasn't. So, we call it explicitly in our eventFilter.

Code cleanup.

Ensure that the indicator disappears on drop events.

Removed some unused code.

* Add the ability for a plug-in to perform custom graphic layer replication.

* Added an AnnotationLayer method to clone an existing GraphicObject into the layer

* Prevent duplicate, ghosty/offset display of icon and text.

* Change PlotView::deletedObjects to take an unordered_set rather than a vector for performance purposes.

* Fixed typo in patch

* More typo fixes.

* Fix remaining plugins except collada.

New issue created to fix collada (447)
New issue created to fix data fusion/flicker tools (448)

+++ Code/application/PlugIns/src/Fits/FitsImporter.cpp
@@ -930,9 +930,14 @@ CachedPage::UnitPtr FitsRasterPager::fetchUnit(DataRequest* pOriginalRequest)
    long maxInRow = std::min(pOriginalRequest->getConcurrentRows(),
       pDesc->getRowCount() - pOriginalRequest->getStartRow().getOnDiskNumber());
    long pixcnt = maxInRow * pDesc->getColumnCount();
+//goffena below
+//   long pStartPix[3] = { 1,
+//                        pOriginalRequest->getStartRow().getOnDiskNumber() + 1,
+//                        pOriginalRequest->getStartBand().getOnDiskNumber() + 1 };
    long pStartPix[3] = {1,
-                        pOriginalRequest->getStartRow().getOnDiskNumber() + 1,
-                        pOriginalRequest->getStartBand().getOnDiskNumber() + 1};
+                        static_cast<long>(pOriginalRequest->getStartRow().getOnDiskNumber() + 1),
+                        static_cast<long>(pOriginalRequest->getStartBand().getOnDiskNumber() + 1)};
+//goffena above
    int status = 0;
 
    size_t bufsize = pixcnt * pDesc->getBytesPerElement();
